Fifistudy/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.4
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
